PlainMem is only for testing
Move all the PlainMem code to a test module, some benches now use
DynamicMem.
