+++ firewood/benches/hashops.rs
@@ -6,12 +6,11 @@
 use criterion::{criterion_group, criterion_main, profiler::Profiler, BatchSize, Criterion};
 use firewood::{
     db::{BatchOp, DbConfig},
-    merkle::{MerkleWithEncoder, TrieHash, TRIE_HASH_LEN},
+    merkle::{TrieHash, TRIE_HASH_LEN, Merkle, Node, Bincode},
     shale::{
-        cached::PlainMem,
         compact::{CompactHeader, CompactSpace},
         disk_address::DiskAddress,
-        CachedStore, ObjCache, Storable, StoredView,
+        ObjCache, Storable, StoredView, cached::DynamicMem, CachedStore,
     },
     storage::WalConfig,
     v2::api::{Db, Proposal},
@@ -19,6 +18,8 @@ use firewood::{
 use pprof::ProfilerGuard;
 use rand::{distributions::Alphanumeric, rngs::StdRng, Rng, SeedableRng};
 use std::{fs::File, iter::repeat_with, os::raw::c_int, path::Path, sync::Arc};
+
+pub type MerkleWithEncoder = Merkle<CompactSpace<Node, DynamicMem>, Bincode>;
 
 const ZERO_HASH: TrieHash = TrieHash([0u8; TRIE_HASH_LEN]);
 
@@ -65,7 +66,7 @@ impl Profiler for FlamegraphProfiler {
 
 fn bench_trie_hash(criterion: &mut Criterion) {
     let mut to = [1u8; TRIE_HASH_LEN];
-    let mut store = PlainMem::new(TRIE_HASH_LEN as u64, 0u8);
+    let mut store = DynamicMem::new(TRIE_HASH_LEN as u64, 0u8);
     store.write(0, &*ZERO_HASH);
 
     #[allow(clippy::unwrap_used)]
@@ -94,7 +95,7 @@ fn bench_merkle<const N: usize>(criterion: &mut Criterion) {
 
                     #[allow(clippy::unwrap_used)]
                     let merkle_payload_header_ref = StoredView::ptr_to_obj(
-                        &PlainMem::new(2 * CompactHeader::MSIZE, 9),
+                        &DynamicMem::new(2 * CompactHeader::MSIZE, 9),
                         merkle_payload_header,
                         CompactHeader::MSIZE,
                     )
@@ -102,8 +103,8 @@ fn bench_merkle<const N: usize>(criterion: &mut Criterion) {
 
                     #[allow(clippy::unwrap_used)]
                     let store = CompactSpace::new(
-                        PlainMem::new(TEST_MEM_SIZE, 0),
-                        PlainMem::new(TEST_MEM_SIZE, 1),
+                        DynamicMem::new(TEST_MEM_SIZE, 0),
+                        DynamicMem::new(TEST_MEM_SIZE, 1),
                         merkle_payload_header_ref,
                         ObjCache::new(1 << 20),
                         4096,
